requirements.txt
aniso8601==8.0.0
attrs==20.2.0
Babel==2.8.0
click==7.1.2
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.4
freeze==3.0
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mysqlclient==2.0.1
PyJWT==1.7.1
pyrsistent==0.17.3
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
Werkzeug==0.16.1
zipp==3.3.0
